[start-coords-cleanup] Consolidate per-component start-coords CSS modules into shared module
Three near-identical .module.css files (exponential, logarithm, tangent)
collapsed into one start-coords-shared.module.css. Adds a .tileRow
variant to capture tangent's flex-row tile layout (vs. the column .tile
used by exponential/logarithm).
