--- packages/perseus-editor/src/widgets/interactive-graph-editor/start-coords/start-coords-exponential.module.css
@@ -1,20 +1,0 @@
-/* We have to use !important until wonder blocks is in the shared layer. */
-/* TODO(LEMS-3686): Remove the !important once we don't need it anymore. */
-.tile {
-    background-color: var(
-        --wb-semanticColor-core-background-instructive-subtle
-    ) !important;
-    margin-top: var(--wb-sizing-size_080) !important;
-    padding: var(--wb-sizing-size_120) !important;
-    border-radius: var(--wb-sizing-size_080) !important;
-}
-
-.row {
-    display: flex !important;
-    flex-direction: row !important;
-    align-items: center !important;
-}
-
-.textFieldWrapper {
-    width: var(--wb-sizing-size_800) !important;
-}
