--- packages/perseus-editor/src/widgets/interactive-graph-editor/start-coords/start-coords-logarithm.module.css
@@ -1,36 +1,0 @@
-/* We have to use !important until wonder blocks is in the shared layer. */
-/* TODO(LEMS-3686): Remove the !important once we don't need it anymore. */
-.tile {
-    background-color: var(
-        --wb-semanticColor-core-background-instructive-subtle
-    ) !important;
-    margin-top: var(--wb-sizing-size_080) !important;
-    padding: var(--wb-sizing-size_120) !important;
-    border-radius: var(--wb-sizing-size_080) !important;
-}
-
-.row {
-    display: flex !important;
-    flex-direction: row !important;
-    align-items: center !important;
-}
-
-.textFieldWrapper {
-    width: var(--wb-sizing-size_800) !important;
-}
-
-.equationSection {
-    margin-top: var(--wb-sizing-size_120) !important;
-}
-
-.equationBody {
-    background-color: var(
-        --wb-semanticColor-core-background-neutral-subtle
-    ) !important;
-    border: var(--wb-border-width-thin) solid
-        var(--wb-semanticColor-core-border-neutral-subtle) !important;
-    margin-top: var(--wb-sizing-size_080) !important;
-    padding-left: var(--wb-sizing-size_080) !important;
-    padding-right: var(--wb-sizing-size_080) !important;
-    font-size: var(--wb-font-size-xSmall) !important;
-}
