+++ packages/perseus-editor/src/widgets/interactive-graph-editor/start-coords/start-coords-logarithm.tsx
@@ -7,7 +7,7 @@ import * as React from "react";
 import CoordinatePairInput from "../../../components/coordinate-pair-input";
 import ScrolllessNumberTextField from "../../../components/scrollless-number-text-field";
 
-import styles from "./start-coords-logarithm.module.css";
+import styles from "./start-coords-shared.module.css";
 import {getLogarithmEquation} from "./util";
 
 import type {Coord} from "@khanacademy/perseus";
